+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [0.0.30] - 2018-??-??
 - (issue 113) Added a command to generate the stub of an HTTP command for the Lytics API. The stub opens in a new editor. The command is available from the command palette or on the context menu for each account listed in the account explorer.
+- (issue 119) Added a command to generate a SipHash. Lytics uses SipHash to convert URLs. The command is available from the command palette.
 
 ## [0.0.29] - 2018-10-08
 - (issue 106) Added a check to the "lytics watch" feature to display an error message if node.js is not installed.